fix: move filtering ignored issues to the LS [IDE-899] (#348)
* fix: move filtering ignored issues to the LS

* fix: improve issue view options texts

Revamp of when the text is shown and the messages themselves, to work with the move of filtering to the LS.

* fix: support LS ignoring IVOs when CCI disabled

The LS will now ignore the Issue View Options when Code Consistent Ignores is not enabled in the org.
As such, the tree nodes have been updated to not display anything about IVOs when CCI is not enabled.

* fix: use constants for messages

Fix some grammar mistakes.
Simplify the logic for the issues found text.

* fix: dont set severity filter in LS init

* fix: show ignores for code quality

* fix: don't return null strings

Added clarifying comments to some issue count variables

* fix: wording changes for issue counts

* fix: make ignored issues hidden by default